requirements.txt
google-api-python-client>=2.88.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
google-auth>=2.28.0
requests>=2.32.0
requests-oauthlib>=2.0.0
yt-dlp>=2025.10.01
python-dotenv>=1.0.0
ffmpeg-python 
pytz>=2024.1
fastapi>=0.110.0
uvicorn>=0.27.0
APScheduler>=3.10.4
musicbrainzngs>=0.7.1
pyacoustid>=1.3.0
mutagen>=1.47.0
rapidfuzz>=3.8.1
Pillow>=10.4.0
